refactor: 💡 modified field and field data type relationship
This commit will modify field and field data type relationship
