+++ src/modules/form/form.module.ts
@@ -4,7 +4,7 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { FieldGroupService } from '../maintenance/field-group/services/field-group.service';
 import { FieldGroupSetService } from './services/field-groupset.service';
 import { FieldInputTypeService } from '../maintenance/field-input-type/services/field-input-type.service';
-import { FieldDataTypeService } from './services/field-data-type.service';
+import { FieldDataTypeService } from '../maintenance/field-data-type/services/field-datatype.service';
 import { FieldOptionService } from '../maintenance/field-option/services/field-option.service';
 import { FieldOptionChildrenService } from './services/field-option-children.service';
 import { FieldOptionGroupService } from './services/field-option-group.service';
@@ -16,7 +16,7 @@ import { FormFieldMemberService } from './services/form-field-member.service';
 import { FormSectionService } from './services/form-section.service';
 import { FormVisibleFieldService } from './services/form-visible-fields.service';
 import { FormSectionFieldMemberService } from './services/formsection-fieldmembers.service';
-import { FieldDatatypeController } from './controllers/field-datatype.controller';
+import { FieldDatatypeController } from '../maintenance/field-data-type/controllers/field-datatype.controller';
 import { FieldGroupController } from '../maintenance/field-group/controllers/field-group.controller';
 import { FieldGroupSetController } from './controllers/field-groupset.controller';
 import { FieldInputTypeController } from '../maintenance/field-input-type/controllers/field-input-type.controller';
@@ -31,7 +31,7 @@ import { FormSectionController } from './controllers/form-section.controller';
 import { FormVisibleFieldController } from './controllers/form-visible-fields.controller';
 import { FormController } from './controllers/form.controller';
 import { FormSectionFieldMemberController } from './controllers/formsection-fieldmembers.controller';
-import { FieldDataType } from './entities/field-datatype.entity';
+import { FieldDataType } from '../maintenance/field-data-type/entities/field-datatype.entity';
 import { FieldGroup } from '../maintenance/field-group/entities/field-group.entity';
 import { FieldGroupSet } from './entities/field-groupset.entity';
 import { FieldInputType } from '../maintenance/field-input-type/entities/field-input-type.entity';
