+++ src/modules/maintenance/field-data-type/controllers/field-datatype.controller.ts
@@ -1,12 +1,10 @@
 import { Controller } from '@nestjs/common';
 import { BaseController } from 'src/core/controllers/base.contoller';
 import { FieldDataType } from '../entities/field-datatype.entity';
-import { FieldDataTypeService } from '../services/field-data-type.service';
+import { FieldDataTypeService } from '../services/field-datatype.service';
 
 @Controller('api/' + FieldDataType.plural)
-export class FieldDatatypeController extends BaseController<
-    FieldDataType
-> {
+export class FieldDatatypeController extends BaseController<FieldDataType> {
     constructor(fieldDataTypeService: FieldDataTypeService) {
         super(fieldDataTypeService, FieldDataType);
     }
