+++ src/modules/maintenance/field-data-type/entities/field-datatype.entity.ts
@@ -1,21 +1,18 @@
 import { Column, Entity, OneToMany } from 'typeorm';
 
-import { Field } from '../../maintenance/field/entities/field.entity';
+import { Field } from '../../field/entities/field.entity';
 import { EntityCoreProps } from 'src/core/entities/entity-core-props';
-
 
 @Entity('fielddatatype', { schema: 'public' })
 export class FieldDataType extends EntityCoreProps {
-
   static plural = 'fieldDataTypes';
 
-  @Column('integer', {
-    nullable: false,
-    primary: true,
-    name: 'fielddatatypeid',
+  /**
+   * One To Many Relationship: FieldDataType and Field
+   */
+  @OneToMany(type => Field, field => field.dataType, {
+    onDelete: 'CASCADE',
+    onUpdate: 'CASCADE',
   })
-  id: number;
-
-  @OneToMany(type => Field, field => field.dataType, { onDelete: 'CASCADE' })
   fields: Field[];
 }
